Resource mining (#27)
* spawning npcs

* handing spawn/despawn of rabbits

* rabbits are now fully functional

* fixed name popup

* Removed updates from individual chunks and parented resources to chunk

* renamed file

* first pass

* updated mouse control scheme so it doesn't clamp with screen edges

* completed control scheme

* Squashed commit of the following:

commit 8ce7155a1e56ff18fbd851ce9a4e603ac85580c6
Author: Alexandre Gauthier <87344030+agauthier5791@users.noreply.github.com>
Date:   Thu Sep 22 14:54:33 2022 -0300

    Npcs (#23)

    * spawning npcs

    * handing spawn/despawn of rabbits

    * rabbits are now fully functional

    * fixed name popup

    * Removed updates from individual chunks and parented resources to chunk

    * renamed file

commit c2906a5433f30ec64862ed30d3f0dde7344af178
Author: Tyler Cloutier <cloutiertyler@users.noreply.github.com>
Date:   Tue Sep 20 23:36:32 2022 -0700

    Implemented changes for compatibility with identity-refactor of spacetimedb (#24)

* Fixed control vs UIs

* colliders and detection

* animations and tools

* increased sensitivity by 20 percent

* resource depletion and vfx

* removed temp code

+++ Client/Assets/_Project/Player/PlayerInventoryController.cs
@@ -38,7 +38,7 @@ public class PlayerInventoryController : MonoBehaviour
         {
             Debug.Log($"Starting item: {item.item.name}, amount: {item.amount}");
             Debug.Assert(NetworkPlayer.localPlayerId != null, "NetworkPlayer._localPlayerId != null");
-            Reducer.AddItemToInventory(NetworkPlayer.localPlayerId.Value, item.item.itemId, (uint)pocketIdx, (int)item.amount);
+            Reducer.AddItemToInventory(NetworkPlayer.localPlayerId.Value, item.item.itemId, pocketIdx, (int)item.amount);
             pocketIdx++;
         }
     }
